Add new colors for WaveColor
This fork needs a facelift. And what better way to start than the current four color system? Four colors is probably good enough for the basic to mid level user. For advanced users with large numbers of tracks, this isn't good enough.
So this commit adds 20 colors to Wave Color selection.